Move live activity code to its own file
